+++ packages/fetch/test/request.js
@@ -88,6 +88,59 @@ describe('Request', () => {
 		expect(r2.counter).to.equal(0);
 	});
 
+	it('should throw a TypeError for forbidden methods', () => {
+		// https://fetch.spec.whatwg.org/#methods
+		const forbiddenMethods = [
+			"CONNECT",
+			"TRACE",
+			"TRACK",
+		];
+
+		forbiddenMethods.forEach(method => {
+			try {
+				new Request(base, { method: method.toLowerCase() });
+				expect(true).to.equal(false);
+			} catch (e) {
+				expect(e instanceof TypeError).to.equal(true);
+				expect(e.message).to.equal(`Failed to construct 'Request': '${method.toLowerCase()}' HTTP method is unsupported.`)
+			}
+			try {
+				new Request(base, { method: method.toUpperCase() });
+				expect(true).to.equal(false);
+			} catch (e) {
+				expect(e instanceof TypeError).to.equal(true);
+				expect(e.message).to.equal(`Failed to construct 'Request': '${method.toUpperCase()}' HTTP method is unsupported.`)
+			}
+		});
+	});
+
+	it('should normalize method', () => {
+		// https://fetch.spec.whatwg.org/#methods
+		const shouldUpperCaseMethods = [
+			"DELETE",
+			"GET",
+			"HEAD",
+			"OPTIONS",
+			"POST",
+			"PUT",
+		];
+		const otherMethods = ["PATCH", "CHICKEN"];
+
+		shouldUpperCaseMethods.forEach(method => {
+			const r1 = new Request(base, { method: method.toLowerCase() });
+			expect(r1.method).to.equal(method.toUpperCase());
+			const r2 = new Request(base, { method: method.toUpperCase() });
+			expect(r2.method).to.equal(method.toUpperCase());
+		});
+
+		otherMethods.forEach(method => {
+			const r1 = new Request(base, { method: method.toLowerCase() });
+			expect(r1.method).to.equal(method.toLowerCase());
+			const r2 = new Request(base, { method: method.toUpperCase() });
+			expect(r2.method).to.equal(method.toUpperCase());
+		});
+	});
+
 	it('should override signal on derived Request instances', () => {
 		const parentAbortController = new AbortController();
 		const derivedAbortController = new AbortController();
@@ -385,5 +438,53 @@ describe('Request', () => {
 			expect(await file.text()).to.equal("content");
 			expect(file.lastModified).to.be.a('number');
 		});
+	});
+
+	it('should decode empty file inputs into File instances (web FormData)', async () => {
+		const ogFormData = new WebFormData();
+		ogFormData.append('a', 1);
+		// This is what happens when you construct the form data set with an empty file input:
+		// https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#constructing-the-form-data-set
+		ogFormData.append('file', new File([], '', { type: 'application/octet-stream' }));
+		const request = new Request(base, {
+			method: 'POST',
+			body: ogFormData,
+		});
+		const clonedRequest = request.clone();
+		return clonedRequest.formData().then(async clonedFormData => {
+			expect(clonedFormData.get('a')).to.equal("1");
+			const file = clonedFormData.get('file');
+			expect(file.name).to.equal("");
+			expect(file.type).to.equal("application/octet-stream");
+			expect(file.size).to.equal(0);
+		});
+	});
+
+	it.skip('should decode empty file inputs into File instances (node FormData)', async () => {
+		const ogFormData = new FormData();
+		ogFormData.append('a', 1);
+		// This is what happens when you construct the form data set with an empty file input:
+		// https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#constructing-the-form-data-set
+		ogFormData.append('file', Buffer.from(''), {
+			// Note: This doesn't work at the moment due to https://github.com/form-data/form-data/issues/412.
+			// There is a v4 released which has a fix that might handle this but I
+			// wasn't positive if it had breaking changes that would impact us so we
+			// can handle an upgrade separately.
+			filename: '',
+			contentType: 'application/octet-stream',
+		});
+		const request = new Request(base, {
+			method: 'POST',
+			body: ogFormData,
+		});
+		const clonedRequest = request.clone();
+		return clonedRequest.formData().then(async clonedFormData => {
+			expect(clonedFormData.get('a')).to.equal("1");
+			const file = clonedFormData.get('file');
+			expect(file.name).to.equal("");
+			expect(file.type).to.equal("application/octet-stream");
+			expect(file.size).to.equal(0);
+		});
+
 	});
 });